% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{airy_a}
\alias{airy_a}
\title{The AiryA Function}
\usage{
airy_a(z, deriv = 0, expon_scaled = FALSE, verbose = 0)
}
\arguments{
\item{z}{A numeric or complex vector representing the input values at which to evaluate the Airy function.}

\item{deriv}{An integer indicating whether to compute the function (\code{0} for the function itself) or its first derivative (\code{1} for the first derivative). Defaults to \code{0}.}

\item{expon_scaled}{A logical value indicating whether to use the exponentially scaled form of the Airy function. Defaults to \code{FALSE}.}

\item{verbose}{An integer specifying the verbosity level for error messages. Defaults to \code{0}.}
}
\value{
A numeric or complex vector (depending on the input) containing the values of the \code{airy_a} function evaluated at the points in \code{z}.
}
\description{
Computes the Airy function Ai for real or complex inputs.
}
\references{
{
\insertRef{Maechler2024}{RcppBessel}\cr
\insertRef{Amos1995}{RcppBessel}
}
}
