% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppBessel-package.R
\docType{package}
\name{RcppBessel-package}
\alias{RcppBessel}
\alias{RcppBessel-package}
\title{RcppBessel: Bessel Functions Rcpp Interface}
\description{
Exports an 'Rcpp' interface for the Bessel functions in the 'Bessel' package, which can then be called from the 'C++' code of other packages. For the original 'Fortran' implementation of these functions see Amos (1995) \doi{10.1145/212066.212078}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alexiosg/RcppBessel}
}

}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} (\href{https://orcid.org/0009-0000-9308-0457}{ORCID})

Authors:
\itemize{
  \item Martin Maechler (\href{https://orcid.org/0000-0002-8685-9910}{ORCID}) (Author of the Bessel R package)
  \item Donald E. Amos (Original author of the zbsubs Fortran code, Sandia National Laboratories)
}

}
\keyword{internal}
