% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppBessel-package.R
\docType{package}
\name{RcppBessel-package}
\alias{RcppBessel}
\alias{RcppBessel-package}
\title{RcppBessel: Bessel C++ headers}
\description{
Exports C++ headers for Bessel functions based on the `Bessel` package.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com}

Authors:
\itemize{
  \item Martin Maechler (Author of the Bessel R package)
  \item Donald E. Amos (Original author of the zbsubs Fortran code, Sandia National Laboratories)
}

}
\keyword{internal}
